+++ database/init.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS containers (
     ip VARCHAR(31) NOT NULL,
     status VARCHAR(31) NOT NULL,
     last_ping_time TIMESTAMP NOT NULL,
-    ping_duration INTERVAL NOT NULL
+    ping_duration BIGINT
 );
 
 -- CREATE TABLE IF NOT EXISTS container_ping_history (
@@ -12,5 +12,5 @@ CREATE TABLE IF NOT EXISTS containers (
 --     ip VARCHAR(31) NOT NULL,
 --     status VARCHAR(31) NOT NULL,
 --     ping_time TIMESTAMP NOT NULL,
---     ping_duration INTERVAL NOT NULL
+--     ping_duration BIGINT
 -- )